+++ src/test/java/com/purbon/kafka/topology/TopologySerdesTest.java
@@ -9,12 +9,13 @@ import com.purbon.kafka.topology.model.Impl.TopologyImpl;
 import com.purbon.kafka.topology.model.Project;
 import com.purbon.kafka.topology.model.Topic;
 import com.purbon.kafka.topology.model.Topology;
+import com.purbon.kafka.topology.model.User;
 import com.purbon.kafka.topology.model.users.Connector;
 import com.purbon.kafka.topology.model.users.Consumer;
-import com.purbon.kafka.topology.model.users.ControlCenter;
 import com.purbon.kafka.topology.model.users.KStream;
 import com.purbon.kafka.topology.model.users.Producer;
-import com.purbon.kafka.topology.model.users.SchemaRegistry;
+import com.purbon.kafka.topology.model.users.platform.ControlCenterInstance;
+import com.purbon.kafka.topology.model.users.platform.SchemaRegistryInstance;
 import com.purbon.kafka.topology.serdes.TopologySerdes;
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -25,6 +26,10 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -185,14 +190,15 @@ public class TopologySerdesTest {
 
     Topology topology = parser.deserialise(Paths.get(topologyDescriptor.toURI()).toFile());
 
-    List<SchemaRegistry> listOfSR = topology.getPlatform().getSchemaRegistry();
+    List<SchemaRegistryInstance> listOfSR =
+        topology.getPlatform().getSchemaRegistry().getInstances();
     assertEquals(2, listOfSR.size());
     assertEquals("User:SchemaRegistry01", listOfSR.get(0).getPrincipal());
     assertEquals("foo", listOfSR.get(0).topicString());
     assertEquals("bar", listOfSR.get(0).groupString());
     assertEquals("User:SchemaRegistry02", listOfSR.get(1).getPrincipal());
 
-    List<ControlCenter> listOfC3 = topology.getPlatform().getControlCenter();
+    List<ControlCenterInstance> listOfC3 = topology.getPlatform().getControlCenter().getInstances();
 
     assertEquals(1, listOfC3.size());
     assertEquals("User:ControlCenter", listOfC3.get(0).getPrincipal());
@@ -209,6 +215,40 @@ public class TopologySerdesTest {
     assertTrue(topology.getProjects().get(0).getProducers().isEmpty());
     assertTrue(topology.getProjects().get(0).getStreams().isEmpty());
     assertTrue(topology.getProjects().get(0).getZookeepers().isEmpty());
+  }
+
+  @Test
+  public void testWithRBACDescriptor() throws IOException, URISyntaxException {
+    URL descriptor = getClass().getResource("/descriptor-with-rbac.yaml");
+
+    Topology topology = parser.deserialise(Paths.get(descriptor.toURI()).toFile());
+    Project myProject = topology.getProjects().get(0);
+
+    assertEquals(2, myProject.getSchemas().size());
+    assertEquals("User:App0", myProject.getSchemas().get(0).getPrincipal());
+    assertEquals(1, myProject.getSchemas().get(0).getSubjects().size());
+
+    Connector connector = myProject.getConnectors().get(0);
+    assertEquals(true, connector.getConnectors().isPresent());
+    assertEquals("jdbc-sync", connector.getConnectors().get().get(0));
+    assertEquals("ibmmq-source", connector.getConnectors().get().get(1));
+
+    Optional<Map<String, List<User>>> rbacOptional =
+        topology.getPlatform().getSchemaRegistry().getRbac();
+    assertTrue(rbacOptional.isPresent());
+
+    Set<String> keys = Arrays.asList("Operator").stream().collect(Collectors.toSet());
+    assertEquals(keys, rbacOptional.get().keySet());
+    assertEquals(2, rbacOptional.get().get("Operator").size());
+
+    Optional<Map<String, List<User>>> kafkaRbacOptional =
+        topology.getPlatform().getKafka().getRbac();
+    assertTrue(kafkaRbacOptional.isPresent());
+
+    Set<String> kafkaKeys =
+        Arrays.asList("SecurityAdmin", "ClusterAdmin").stream().collect(Collectors.toSet());
+    assertEquals(kafkaKeys, kafkaRbacOptional.get().keySet());
+    assertEquals(1, kafkaRbacOptional.get().get("SecurityAdmin").size());
   }
 
   private List<Project> buildProjects() {