+++ src/test/java/com/purbon/kafka/topology/integration/MDSApiClientRbacIT.java
@@ -1,16 +1,21 @@
 package com.purbon.kafka.topology.integration;
 
 import static com.purbon.kafka.topology.roles.RBACPredefinedRoles.DEVELOPER_READ;
+import static com.purbon.kafka.topology.roles.RBACPredefinedRoles.RESOURCE_OWNER;
+import static com.purbon.kafka.topology.roles.RBACPredefinedRoles.SECURITY_ADMIN;
 import static com.purbon.kafka.topology.roles.RBACProvider.LITERAL;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 import com.purbon.kafka.topology.AccessControlManager;
 import com.purbon.kafka.topology.api.mds.AuthenticationCredentials;
 import com.purbon.kafka.topology.api.mds.MDSApiClient;
 import com.purbon.kafka.topology.roles.RBACProvider;
+import com.purbon.kafka.topology.roles.TopologyAclBinding;
 import java.io.IOException;
 import java.util.List;
+import java.util.Map;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.Before;
@@ -70,5 +75,47 @@ public class MDSApiClientRbacIT extends MDSBaseTest {
     List<String> roles = apiClient.lookupRoles("User:fry");
     assertEquals(1, roles.size());
     assertTrue(roles.contains(DEVELOPER_READ));
+  }
+
+  @Test
+  public void testBindRoleWithoutAuthentication() throws IOException {
+    apiClient.setKafkaClusterId(getKafkaClusterID());
+
+    TopologyAclBinding binding =
+        apiClient.bind("User:fry", DEVELOPER_READ, "connect-configs", LITERAL);
+
+    assertNull(binding);
+  }
+
+  @Test
+  public void testBindSecurityAdminRole() throws IOException {
+    apiClient.login(mdsUser, mdsPassword);
+    apiClient.authenticate();
+    apiClient.setKafkaClusterId(getKafkaClusterID());
+    apiClient.setSchemaRegistryClusterID("schema-registry");
+    String principal = "User:foo" + System.currentTimeMillis();
+
+    apiClient.bind(principal, SECURITY_ADMIN).forSchemaRegistry().apply();
+
+    Map<String, Map<String, String>> clusters =
+        apiClient.withClusterIDs().forKafka().forSchemaRegistry().asMap();
+
+    List<String> roles = apiClient.lookupRoles(principal, clusters);
+    assertEquals(1, roles.size());
+    assertTrue(roles.contains(SECURITY_ADMIN));
+  }
+
+  @Test
+  public void testBindResourceOwnerRole() throws IOException {
+    apiClient.login(mdsUser, mdsPassword);
+    apiClient.authenticate();
+    apiClient.setKafkaClusterId(getKafkaClusterID());
+
+    String principal = "User:fry" + System.currentTimeMillis();
+    apiClient.bind(principal, RESOURCE_OWNER, "connect-configs", LITERAL);
+
+    List<String> roles = apiClient.lookupRoles(principal);
+    assertEquals(1, roles.size());
+    assertTrue(roles.contains(RESOURCE_OWNER));
   }
 }